refactor: 관리자 페이지 타입 시스템 정비 및 컴포넌트 분리
- types/admin.ts 생성으로 중앙 집중식 타입 관리
- any 타입 제거 및 타입 안전성 향상
- CouponManagement 컴포넌트 별도 파일로 분리
- import type 활용으로 번들 최적화
- 매직 스트링을 상수로 교체하여 유지보수성 개선
